fix: Stop adding unwanted newlines in multiline strings
This patch set fixes an tools/updater.py YAML export issue in which pyyaml is
adding unwanted (multiple) newlines in multiline strings in the output. In
other words, currently, tools/updater.py keeps injecting additional newline
characters in multiline certificates, such that each newline is rendered, in
effect, as two newlines.

For example:

  -----BEGIN CERTIFICATE-----
  MIIDSDCCAjCgAwIBAgIUegkh/antB1XyDVHdP5dv+0MZyBcwDQYJKoZIhvcNAQEL
  BQAwKjETMBEGA1UEChMKS3ViZXJuZXRlczETMBEGA1UEAxMKa3ViZXJuZXRlczAe
  <snip>

Currently renders as:

  -----BEGIN CERTIFICATE-----

  MIIDSDCCAjCgAwIBAgIUegkh/antB1XyDVHdP5dv+0MZyBcwDQYJKoZIhvcNAQEL

  BQAwKjETMBEGA1UEChMKS3ViZXJuZXRlczETMBEGA1UEAxMKa3ViZXJuZXRlczAe
  <snip>

This patch set adds code to the updater.py tool to use the
solution prescribed in [0] to force pyyaml to ensure that the
rendered output for such multiline strings matches the original
input.

Change-Id: I8b5c5166a0a1b3d945281addd4f562368e1907c3