global: Add labels to OSH charts for Airskiff site
The Airskiff site needs the ability to override charts that have
multiple replicas or utilize Prometheus. This commit adds labels to
OpenStack-Helm charts and chart groups that will be overriden in the
Airskiff site. Charts are labeled with name and component selectors to
match existing labels on select charts and leave the `airship-seaworthy`
site unaffected.

Change-Id: I24d69afe70fbee35d3b21b8803eb64a81ed5376e