airskiff: Pin Armada to K8s v2.11.0
Currently, OpenStack-Helm-Infra uses version 2.11.0 of Kubernetes.
Because Airskiff uses OpenStack-Helm-Infra for deployment of K8s, newer
versions of Armada cause the Airskiff gate to fail, as Armada now uses
Tiller v2.12.1. This commit pins Armada to the latest version that
utilized Tiller v2.11.0. This pin should be removed as soon as
OpenStack-Helm-Infra moves to v2.12.1.

Change-Id: Iaa3a776ffda9678b7736c1b1aac72a1adbedf29b