airskiff: Update Armada image to pinned version
Previously, Armada was pinned [0] to commit
dd75474d78e69f3634eb94c8ea6730b898e83277 in order to prevent a mismatch
between the Helm API and Tiller. While the generated image tag was
updated, the Armada image deployed by Armada during the bootstrapping
phase was not updated, resulting in a version mismatch. This commit
overrides the tag in order to prevent a mismatch.

[0] https://review.openstack.org/630698

Change-Id: I484c5e52560e5c25c76ad557f3032ae9e158166c