airskiff: Disable rabbitmq tests
Currently, the tests for the OSH rabbitmq charts are disabled; however,
the value that disables the chart is sometimes discarded during document
layering because the `.test` value does not get merged with the parent
document.  This commit adds an action to merge the test key and disable
the rabbitmq tests.

Change-Id: Ia23042bd03871cc085f69c232a85375571282af6