airskiff: Add steps to Shipyard wait script
Currently, the Airskiff gate does not recognize failures because the
gate script only verifies that an action lifecycle enters a COMPLETE
state; however, a lifecycle in COMPLETE status is not always an
indication of a successful deployment. This commit updates the
wait_for_shipyard script to check the status of each action step and
results in exit_code 1 if any steps failed.

Change-Id: If5117c4638652b8d5b91fe73ede0f5e19a1e2cc1