airskiff: Update OSH-infra pin
The version of Armada deployed by Airskiff was recently uplifted. The
gate erroneously uses the latest version of OSH-infra, so there was no
impact; however, users of the 005-make-airship script experience a
Tiller version mismatch due to an older version of K8s. This change
uplifts the OSH-infra pin to prevent errors.

Change-Id: If8f2c27396ce6c8548598492dfaf5e15174131a2