Update from docker-engine to docker-ce
A previous patch that bumps the version of docker [1]
is broken because there is no "17.03.3" version of
docker-engine.

Docker package is now called docker-ce [2] and is provided
from a different repo, update repo paths and keys
to newer docker version.

The key is taken from: https://download.docker.com/linux/ubuntu/gpg

[1] https://review.openstack.org/#/c/639175/
[2] https://docs.docker.com/install/linux/docker-ce/ubuntu/

Change-Id: Idddee15e17a75017fdb0b8401aafef86cec4af1f