docs: Add initial contributors' guide
This change introduces an initial contributors' guide, with the
intention of providing a starting point for developers interested in
contributing to Airship.

Future changes should expand on this documentation, including the
addition of documentation that covers the spec design and approval
process, standards and conventions for Helm charts, and how versions are
controlled in Airship.

Change-Id: Ifd146549d0707f9db2aefaefbadc77e47a1546e3
Story: 2005122