Airskiff: update packages prior to delete
Previously, airskiff's 000-install-packages script added the docker
repo, then deleted stale docker packages, then updated the apt database.
This change moves the apt update before the delete, so that apt-get
knows about the packages it's being asked to delete.  Without this in
place it'll fail on a fresh VM that doesn't know about the docker repo.

Change-Id: If009d8371c3ff49e4ed50599d8d62072868a84bf