Increase default logging for openstack services
Globals have logging override defined which is now different from
openstack services values in openstack-helm.

Openstack-helm addressed issue of missing logs related to warnings and
above for root logger which is catch-all logger when component specific
logger is not explictly defined.

OSH review: https://review.opendev.org/#/c/639711

Dropping logs in case of error or warnings can be problem for operators
in triaging an issue. We should not be hiding warning/error/critical logs
even when specific component logger is not defined as there can be cases
where each component is not defined in logging definition. See example
from OSH review.

In addition, is there a specific reason to have logging override for OS services
in airship side as they appear to be quite similar to defaults defined in
openstack-helm side. So may be we should revisit its need and address in
different patch if needed.

Change-Id: Iea24280546af1959d9da7164f0f18bf6a0a86e2d