airskiff: Use Ubuntu images
Since support for additional operating systems has been added to some
components, the image tag "latest" is no longer used by their Makefiles.
This commit changes the tags for the aforementioned components to use
the new default tag and updates the job name to reflect the use of
Ubuntu images.

Change-Id: Ia1adf156c4b3798b2400f9c821ac2ba685557591