030-armada-boostrap.sh: make --target-manifest env var driven
Currently the manifest targeted by armada in the
030-armada-bootstrap.sh script is hard-coded to "ucp-bootstrap". This
patch set updates the 030 script to make drive the target manifest
off of a new environment variable "TARGET_MANIFEST". By default
"ucp-bootstrap" is still used, and existing functionality should not
be altered.
The motivation behind this, is it allows us to use the armada
bootstrap script to do more than just bootstrap the ucp components,
which may prove useful in zuul gate jobs that use treasuremap.

Change-Id: If3cf11f0d5701e273f5e626bff673a2b164284e3