Introduce CoreDNS DaemonSet
This enables coredns to run as a daemonset across the control plane
hosts, for increased resiliency.  This change retains the deployment-
controlled coredns pods as well (with two replicas) to make sure there
is no interruption to DNS service while the change is rolled out in
existing Airship deployments.  The deployment can be turned off in a
future release, once the daemonset is in place.

Change-Id: I4678ff2490ab36a2f5a1cc5030492d85ef70b7ac