Bumped the timeout for the ucp keystone test
Was brought up on the last AirShip IRC meeting @ 20.08.2019

When deploying AIAB, it is possible that the ucp keystone test will exceed the deadline,
and this will happen on each retry - which will cause the test to loop. In order to avoid this,
I've bumped the test timeout in the keystone chart to 600s, as in all my cases, the test used ~371 seconds to complete.

Change-Id: I55d1fa0d14150269a34514d523d5ee764ec057f9