Fix charts overrides for hyperkube 1.12
Promenade chart

The issue is that for joined nodes we will see hyperkube version
from Promenade chart which is not updated or different.

So, need to substitute hyperkube image directly in override to
align all hyperkube version in the cluster.

Tiller chart

Override image for tiller was defined under armada, so, need
to use appropriate place for source in substitution.

Removed placeholder to eliminate duplicates.

Change-Id: I01b8be41591f85fa962c0524bff75f0fdc97100d