Changed Nova compute endpoint to new version
Current Nova service endpoint is set to legacy compute version: v2.
According to the doc [1], minimum microversion in Nova is 2.1,
which is a microversion compatible with the legacy v2 API.
After that, nova client will be able to add into a requests:
X-OpenStack-Nova-API-Version HTTP header, which is required for some
tempest tests.

[1] https://docs.openstack.org/api-guide/compute/microversions.html

Change-Id: I6a7305cbbab0b20a6b926f6358dfdf463df49652