Fix pyyaml load call error
yaml.load(input) is deprecated in pyyaml 5.1. This causes the
manifests_lookup function to append the python warning message
to the look up results. The pegleg cmd fails to launch container
due to extra strings after the image url. Added code to specify
the loader if the pyyaml version is 5.1 or newer.

Change-Id: I4765df28eedf99a8e823e192cf2f2f2b38434f2c
Signed-off-by: James Gu <james.gu@att.com>