Separated Ceph Content into separate section
There is a lot of troubleshooting content related to Ceph that
will be submitted in following patchsets related to
common issues realted to Monitors, OSDs, and PSGs. The amount
of content would overshadow the more surface level instructions
in the troubleshooting guide. So I have given Ceph it's own
separated  section under the Troubleshooting Guide

Change-Id: I27f67f6813eed4823ea5ff43bcadd7d269a8afa9