Add skeleton of airship-core type
This adds a basic airship-core type and test-site definition, which initially
will point back to airshipctl for everything but site-specific configuration.

Gating is added under a dependent changeset:
https://review.opendev.org/#/c/741648

This removes the yaml-lint job / make target, since that chokes on CRs whose
definitions are not present.  YAML validation will be re-enabled by borrowing
the YAML validation from the airshipctl project, which leverages kustomize
and a live kubernetes API server.

Change-Id: I51c469f74be1270325f523343c02d36c2512e739