Add network-cloud type
This change introduces an outline of the network-cloud type, which will
aim to provide Airship configuration suitable for performance-intensive
telecommunication and networking workloads according to the CNTT
specification.

Most documents include in this type inherit from the airship-core type;
however, this type will evolve to include type-specific documents in the
future.

Change-Id: I98344649ca55dc00151734963265b26d28deec68