Add version catalogues
This takes advantage of version catalogue replacement, which was added to
airshipctl in the changes below.  It also refactors the kustomize structure
a bit to support that well.

This adds a treasuremap clusterctl function, which kustomizes the same function
from airshipctl to correct for paths.  This replaces the previous
site-level clusterctl entrypoints.

This gives each type and site a catalogues entrypoint, which are wired together
(site inheriting from type, and network-cloud from airship-core) for
inheritance.  Each phase that needs catalogue-based replacement now
pulls in its' site's catalogues definition, and then applies any
phase specific overrides only when necessary.

https://review.opendev.org/#/c/746608
https://review.opendev.org/#/c/746764/

Change-Id: I18364083f72628f9519af3d2fab0d6a4879d2d14
Relates-To: https://github.com/airshipit/airshipctl/issues/316