Add document pull functionality to deploy job
This change introduces airshipctl's document pull functionality to the
Treasuremap gate runner job. This enables Treasuremap maintainers to pin
to a version of airshipctl deployment scripts and manifests that are
known to be compatible with Treasuremap's v2 branch.

In order to ensure that the deployment scripts and manifests come from
the same version of airshipctl, this change introduces a playbook that
checks out the referenced version of airhsipctl after it is cloned by
Zuul. It also checks out the known version a second time in the
deployment scripts for operators invoking the scripts in a non-Zuul
environment.

To update the pinned version of Treasuremap using Zuul, change the Zuul
variable airshipctl_ref. To update the pinned version in a non-Zuul
environment, use the environment variable AIRSHIPCTL_REF.

Change-Id: I300f545efcdd024b689a96ee99e502de7fdd1b83
Signed-off-by: Drew Walters <andrew.walters@att.com>