Patch hostgenerators to delete catalogues
To facilitate catalogue-driven generation of BareMetalHost resources,
airship splits the work into two pieces:

1. <phasename>/hostgenerator/ is responsible for substituting
   catalogue variables into a host generation template
2. <phasename>/ invokes its hostgenerator/ as a Kustomize generator:,
   leveraging the results of 1. to generate BMH.

However, a weird side-effect of this setup is that the catalogues used
in step 1. above are still hanging around in step 2., and Kustomize
tries to interpret them as generator: plugin configs,
resulting in a rendering error.

Previously, we'd worked around this by having a pass-through plugin
implementation, which performs a no-op generation for the spent catalogues.

However, with the change below, airshipctl migrated to container-based KRM
functions for its kustomize plugins, and also adopted a
strategic-patch-delete mechanism to clean up the spent catalogues,
so that they aren't still hanging around by the time the generation occurs.

https://review.opendev.org/#/c/759279/

Change-Id: I50a9bd2fcc37b70f50b8ded88f152775aff928bc
Closes: #48