Update deployment job timeout
The Treasuremap deployment jobs sometimes timeout prior to the desired,
script-defined timeout of the worker node deployment script. This change
increases the timeout of the deployment jobs to account for the
difference in time that the deployment script specifies.

This will make the jobs more reliable; however, there is still another
deployment error that occurs that will prevent us from making the jobs
voting in the meantime.

Signed-off-by: Drew Walters <andrew.walters@att.com>
Change-Id: I7b1bae7bb4445516ebef11cd7a92730e95512462