Temporarily disable non-voting gates
As we prepare for the v2.0 release, we're trying to merge a large number
of patches, which seems to be stressing out our infrastructure, resulting
in less merging instead of more.

This change changes the (non-voting) integration checks that are running
in citycloud nodepool to experimental checks, with the intent of
reverting this change in a day or two after the mergestorm subsides.

Change-Id: If2e95cb800b6aeac84fe1399687027b2143c5d07