Bump airshipctl pin
This bumps the airshipctl pin to the latest merged commit as of 4/1 a.m.,
which includes the kustomize install script[1]  needed by the corresponding
treasuremap change[2], as well as the helm-chart-collator fix here[3].

It also reverts the treasuremap versions catalogue to a VariableCatalogue
kind, since it doesn't currently adhere to the VersionsCatalogue schema.
That will need to be trued up post-2.0-release.

1: https://review.opendev.org/c/airship/airshipctl/+/783907
2: https://review.opendev.org/c/airship/treasuremap/+/783911
3: https://review.opendev.org/c/airship/airshipctl/+/784122

Change-Id: I40d81c6bbe20a9a24af26cbb50791d6afef50149