Fix incorrect reference-airship-core type refs
Some phase entrypoints were missing in the multi-tenant type when the
baremetal reference-multi-tenant-site was created. When the multi-tenant
type was disassociated from the airship-core type [0], those entrypoints
were created and added to all virtual sites; however, the reference
multi-tenant site was not updated. This change fixes those references in
the reference-multi-tenant site.

[0] https://review.opendev.org/c/airship/treasuremap/+/786888

Signed-off-by: Drew Walters <andrew.walters@att.com>
Change-Id: Icc346615a1a6de9cfd4a93e9b8a89a4cd0a640e0