Fix workload replacements
The workload entrypoint does not have an replacements directory, meaning
that each function's replacements must be listed at the site level.
Relying on calling a function's replacements can be unreliable when
making changes at the type level because entrypoints also have to be
updated at the site level. This change adds a replacements entrypoint
for the workload phases so that changes can be made at the type level
without changing each site.

Signed-off-by: Drew Walters <andrew.walters@att.com>
Change-Id: I450fb5e57967b1fe0b585f85686b558962338d90