Use upstream rook yamls as base
This is a scaled back version of:

https://review.opendev.org/c/airship/treasuremap/+/792837

This PS basically uses kustomize patches to modify the upstream
yamls to create the base resources. Currently we manually
edit the upstream yamls. The yamls produced by "kubectl kustomize ."
are as before.

Kpt can be used for syncing to upstream once this issue is resolved:

https://github.com/GoogleContainerTools/kpt/issues/517

Change-Id: Idd7a3f4a61a9dae9ec82dd0a604eb371bdbb4ab4