Absorb helm chart collator functions
This copies the hcc functions from airshipctl into treasuremap.
They are not used in the airshipctl repo, and thus having them
there doesn't add value, and also led to issues which were
not uncovered until we uplifted airshipctl in treasuremap [0].
This moves them to treasuremap where they are actually used.

They will need to be removed from airshipctl separately.

[0]: https://review.opendev.org/c/airship/airshipctl/+/794707

Signed-off-by: Sean Eagan <seaneagan1@gmail.com>
Change-Id: Ief295eaa9b84b4303add537b3526b425c91014e6