Use kpt for rook
This PS is to enable kpt for rook-ceph.

The situation is odd in that we need to use

https://github.com/rook/rook/tree/master/cluster/examples/kubernetes/ceph

as an upstream for each base directory but only need a single file in
each case.

The solution here is to clone the entire upstream directory into each
base directory. Gitignore files are used to prevent having to commit
hundreds of files to the repo.

This is obviously not very efficient but it is the best I could come
up with.

Hopefully kpt will be able to sync single files on the future.

Change-Id: I2a5fa63b56749653880be72333a7fce1620bdb6c