Configure ssh from jump host to VMs
The SipCluster CRs reference an ssh-private-keys Secret
which did not exist. This is used to ssh into the VMs
from within the Sip jump hosts. This change adds the Secret
and sources its content from the ssh keys generated for the
target cluster. Once we have a pattern for generating secrets
for sub-clusters in place we should generate separate ssh
key pairs for each sub-cluster.

Signed-off-by: Sean Eagan <seaneagan1@gmail.com>
Change-Id: I7eb997e6328bba0fdc49f5b9b5a8c097fc24e143